Rename button to Button for casing consistency
